requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
python-dateutil==2.9.0.post0